requirements.txt
async-generator==1.10
atomicwrites==1.4.1
attrs==22.1.0
awsebcli==3.20.3
botocore==1.23.54
cement==2.8.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.3
dnspython==2.2.1
dominate==2.7.0
email-validator==1.2.1
filelock==3.8.0
Flask==2.2.2
Flask-Bootstrap==3.3.7.1
Flask-WTF==1.0.1
future==0.16.0
h11==0.13.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==0.10.0
MarkupSafe==2.1.1
numpy==1.24.3
outcome==1.2.0
packaging==21.3
pathspec==0.9.0
pluggy==1.0.0
py==1.11.0
py3-validate-email==1.0.5
pycparser==2.21
pyparsing==3.0.9
pypiwin32==223
PySocks==1.7.1
pytest==7.1.2
pytest-flask==1.2.0
pytest-lazy-fixture==0.6.3
pytest-mock==3.8.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pywin32==305
PyYAML==5.4.1
regex==2022.7.25
requests==2.26.0
requests-mock==1.9.3
selenium==4.4.3
semantic-version==2.8.5
six==1.14.0
sniffio==1.2.0
sortedcontainers==2.4.0
termcolor==1.1.0
tomli==2.0.1
trio==0.21.0
trio-websocket==0.9.2
tzdata==2023.3
urllib3==1.26.11
visitor==0.1.3
wcwidth==0.1.9
Werkzeug==2.2.2
wsproto==1.2.0
WTForms==3.0.1
